ITS WORKING
It is by no means efficient but it works and ima call that a win. There was a typo in Dr. Alvin's test cases that let me pass the implicit segments tests where I shouldnt have been able to. After discovering this and creating a isMinimalSegment Method I was able to fix everything and get the magical green light.